requirements.txt
pandas==1.4.1
numpy==1.22.2
scipy==1.8.0
scikit-surprise==1.1.1
recommenders==1.1.0
tf_slim==1.1.0
tensorflow==2.9.1